refactor: extract the public logic for the bucket-type bulk insert partitioner
1. extract the public logic for the bucket-type bulk insert partitioner
